Add a board-info example for old radio firmware versions.
While this is not supported, it lists the radio firmwares that shipped
on Nexus One with builds ESD79 and ERE27, and if someone has one
of those firmwares and can't update it, using this file instead of
the regular board-info.txt will still allow flashing a system newer than
eclair on the device with a plain fastboot flashall.

Of course the result is entirely unsupported. It is even known to be
unstable, to the point where some functions cause the phone to entirely
crash.

Bug: 2581353
Change-Id: I7012996f78575f944df106b47c6fde2c7d9f395b